mover Genero a su propio archivo
